Migrating correlation_id to BigInteger

+++ iaso/models.py
@@ -944,7 +944,7 @@ class Instance(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
     uuid = models.TextField(null=True, blank=True)
     export_id = models.TextField(null=True, blank=True, default=generate_id_for_dhis_2)
-    correlation_id = models.IntegerField(null=True, blank=True)
+    correlation_id = models.BigIntegerField(null=True, blank=True)
     name = models.TextField(null=True, blank=True)
     file = models.FileField(upload_to=UPLOADED_TO, null=True, blank=True)
     file_name = models.TextField(null=True, blank=True)
